Add C3GovernorUpgradeable module

+++ src/gov/C3GovernDapp.sol
@@ -1,45 +1,39 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 pragma solidity ^0.8.19;
 
-import {SafeERC20} from "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
-import {Math} from "@openzeppelin/contracts/utils/math/Math.sol";
-import {Strings} from "@openzeppelin/contracts/utils/Strings.sol";
-import {Address} from "@openzeppelin/contracts/utils/Address.sol";
+import { SafeERC20 } from "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
 
-import {C3CallerDapp} from "../dapp/C3CallerDapp.sol";
-import {C3GovClient} from "../gov/C3GovClient.sol";
-import {TheiaUtils} from "../theia/TheiaUtils.sol";
+import { Address } from "@openzeppelin/contracts/utils/Address.sol";
+import { Strings } from "@openzeppelin/contracts/utils/Strings.sol";
+import { Math } from "@openzeppelin/contracts/utils/math/Math.sol";
+
+import { C3CallerDapp } from "../dapp/C3CallerDapp.sol";
+import { C3GovClient } from "../gov/C3GovClient.sol";
+import { TheiaUtils } from "../theia/TheiaUtils.sol";
 
 abstract contract C3GovernDapp is C3CallerDapp {
     using Strings for *;
     using Address for address;
     // delay for timelock functions
-    uint public delay = 2 days;
+
+    uint256 public delay = 2 days;
 
     address private _oldGov;
     address private _newGov;
     uint256 private _newGovEffectiveTime;
 
-    mapping (address => bool) txSenders;
+    mapping(address => bool) txSenders;
 
-    constructor(
-        address _gov,
-        address _c3callerProxy,
-        address _txSender,
-        uint256 _dappID
-    ) C3CallerDapp(_c3callerProxy, _dappID) {
+    constructor(address _gov, address _c3callerProxy, address _txSender, uint256 _dappID)
+        C3CallerDapp(_c3callerProxy, _dappID)
+    {
         _oldGov = _gov;
         _newGov = _gov;
         _newGovEffectiveTime = block.timestamp;
         txSenders[_txSender] = true;
     }
 
-    event LogChangeGov(
-        address indexed oldGov,
-        address indexed newGov,
-        uint indexed effectiveTime,
-        uint256 chainID
-    );
+    event LogChangeGov(address indexed oldGov, address indexed newGov, uint256 indexed effectiveTime, uint256 chainID);
 
     event LogTxSender(address indexed txSender, bool valid);
 
@@ -60,15 +54,10 @@ abstract contract C3GovernDapp is C3CallerDapp {
         _oldGov = gov();
         _newGov = newGov;
         _newGovEffectiveTime = block.timestamp + delay;
-        emit LogChangeGov(
-            _oldGov,
-            _newGov,
-            _newGovEffectiveTime,
-            block.chainid
-        );
+        emit LogChangeGov(_oldGov, _newGov, _newGovEffectiveTime, block.chainid);
     }
 
-    function setDelay(uint _delay) external onlyGov {
+    function setDelay(uint256 _delay) external onlyGov {
         delay = _delay;
     }
 
@@ -86,19 +75,14 @@ abstract contract C3GovernDapp is C3CallerDapp {
         return txSenders[txSender];
     }
 
-    function doGov(
-        string memory _to,
-        string memory _toChainID,
-        bytes memory _data
-    ) external onlyGov {
+    function doGov(string memory _to, string memory _toChainID, bytes memory _data) external onlyGov {
         c3call(_to, _toChainID, _data);
     }
 
-    function doGovBroadcast(
-        string[] memory _targets,
-        string[] memory _toChainIDs,
-        bytes memory _data
-    ) external onlyGov {
+    function doGovBroadcast(string[] memory _targets, string[] memory _toChainIDs, bytes memory _data)
+        external
+        onlyGov
+    {
         require(_targets.length == _toChainIDs.length, "");
         c3broadcast(_targets, _toChainIDs, _data);
     }
